+++ src/test/java/nextstep/subway/acceptance/StationAcceptanceTest.java
@@ -1,7 +1,7 @@
 package nextstep.subway.acceptance;
 
 import io.restassured.RestAssured;
-import nextstep.subway.applicaion.dto.StationResponse;
+import nextstep.subway.applicaion.dto.response.StationResponse;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
