+++ src/test/java/nextstep/subway/unit/service/LineServiceTest.java
@@ -1,7 +1,7 @@
 package nextstep.subway.unit.service;
 
 import nextstep.subway.applicaion.LineService;
-import nextstep.subway.applicaion.dto.SectionRequest;
+import nextstep.subway.applicaion.dto.request.SectionRequest;
 import nextstep.subway.domain.Line;
 import nextstep.subway.domain.LineRepository;
 import nextstep.subway.domain.Station;
